prompt_engineering/pyproject.toml
[project]
name = "prompt-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.2.0",
    "pillow>=12.1.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
